feat(backup): Generate validation run cloudbuild.yaml
This file instruments the validation run we need to perform before we
can kick off the actual relocation. We build a bespoke YAML for each
such run, rather than using substitutions, as there is enough
variability that this ends up a bit easier to read, and some variables
could grow longer than the substitution API is willing to accept.

Issue: getsentry/team-ospo#203