feat(metric-alerts): Explicitly add entity to columns for error alerts
We want to allow error metric alerts to be filtered by status. To do this, we need to be able to join to the issue level table that has been replicated to Snuba. Snql supports this, but when defining our query we'll need to specify which entity each column belongs to.

This pr does the preliminary work of adding an errors specific `QueryBuilder` that handles adding entities onto all existing columns, and uses it in `BaseEventsAndTransactionEntitySubscription`. Actual work to filter on the status column will be in a follow up.